# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VDMS API",
        "description": "Vehicle Dealership Management System API",
        "contact": {
            "email": "support@vdms.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/vdms/public/api/v1",
            "description": "Development Server"
        },
        {
            "url": "https://waba.insightechindia.in/public/api/v1/",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/v1/system-settings": {
            "get": {
                "tags": [
                    "System Settings"
                ],
                "summary": "List all system settings grouped by topic",
                "description": "Returns all active system settings organized by topic/category",
                "operationId": "indexSettings",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all settings"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/system-settings/topic/{topic}": {
            "get": {
                "tags": [
                    "System Settings"
                ],
                "summary": "Get settings for a specific topic",
                "description": "Returns all settings for the specified topic",
                "operationId": "showTopic",
                "parameters": [
                    {
                        "name": "topic",
                        "in": "path",
                        "description": "Topic/category name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved topic settings"
                    },
                    "404": {
                        "description": "Topic not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/system-settings/key/{key}": {
            "get": {
                "tags": [
                    "System Settings"
                ],
                "summary": "Get a single setting by key",
                "description": "Returns the value of a specific setting",
                "operationId": "showKey",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Setting key to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved setting"
                    },
                    "404": {
                        "description": "Setting not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "System Settings"
                ],
                "summary": "Update a setting value (admin only)",
                "description": "Updates the value of a specific setting",
                "operationId": "updateSetting",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Setting key to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "value"
                                ],
                                "properties": {
                                    "value": {
                                        "type": "string",
                                        "example": "New Value"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Setting updated successfully"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    },
                    "404": {
                        "description": "Setting not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/system-settings/category/site": {
            "get": {
                "tags": [
                    "System Settings"
                ],
                "summary": "Get site configuration settings",
                "description": "Get site-specific settings\n\nConvenience endpoint for site configuration settings.",
                "operationId": "getSiteSettings",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved site settings"
                    },
                    "404": {
                        "description": "Settings not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/system-settings/category/dealership": {
            "get": {
                "tags": [
                    "System Settings"
                ],
                "summary": "Get dealership information settings",
                "description": "Get dealership-specific settings\n\nConvenience endpoint for dealership information settings.",
                "operationId": "getDealershipSettings",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved dealership settings"
                    },
                    "404": {
                        "description": "Settings not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/system-settings/category/pricing": {
            "get": {
                "tags": [
                    "System Settings"
                ],
                "summary": "Get pricing configuration settings",
                "description": "Get pricing-specific settings\n\nConvenience endpoint for pricing configuration settings.",
                "operationId": "getPricingSettings",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved pricing settings"
                    },
                    "404": {
                        "description": "Settings not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/system-settings/export/json": {
            "get": {
                "tags": [
                    "System Settings"
                ],
                "summary": "Export all settings as JSON",
                "description": "Export all settings as JSON\n\nExports all system settings for backup or migration (admin only).",
                "operationId": "exportJson",
                "responses": {
                    "200": {
                        "description": "Settings exported successfully"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/system-settings/import/json": {
            "post": {
                "tags": [
                    "System Settings"
                ],
                "summary": "Import settings from JSON",
                "description": "Import settings from JSON\n\nImports system settings from JSON array (admin only).",
                "operationId": "importJson",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "settings"
                                ],
                                "properties": {
                                    "settings": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "key": {
                                                    "type": "string"
                                                },
                                                "value": {
                                                    "type": "string"
                                                },
                                                "topic": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import completed"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}